+++ ThemeDecorator/ThemeDecorator.js
@@ -256,12 +256,10 @@ const ThemeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 	const {customSkin, float, i18n, noAutoFocus, overlay, ri, skin, spotlight, disableFullscreen} = config;
 	const defaultSkin = 'gallium';
 
-	const bgClassName = classnames(
-		{
-			'enact-fit': !disableFullscreen,
-			[css.bg]: !overlay
-		}
-	);
+	const bgClassName = classnames({
+		'enact-fit': !disableFullscreen,
+		[css.bg]: !overlay
+	});
 
 	let App = Wrapped;
 	if (float) App = FloatingLayerDecorator({wrappedClassName: bgClassName}, App);