+++ CHANGELOG.md
@@ -17,11 +17,11 @@ The following is a curated list of changes in the Enact agate module, newest cha
 - `agate/Popup` content padding and margin with `closeButton`
 - `agate/Popup` `bottom`, `fullscreen`, `left`, and `right` values to `position` prop
 - `agate/Popup` `duration` and `type` props
-- `agate/ThemeDecorator` class `enact-fit` based on the value of `disableFullScreen` prop
 
 ### Changed
 
 - `agate/Heading` to support marquee
+- `agate/ThemeDecorator` to exclude `enact-fit` className when `disableFullscreen` is true
 
 ### Fixed